refactor: mejoras en las carpetas de estructura del proyecto
