+++ lib/core/utils/app_info.dart
