+++ lib/core/utils/battery_utils.dart
