+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
-import 'package:webradiooasis/Pages/church_main_screen.dart';
-import 'package:webradiooasis/services/audio_player_provider.dart';
+import 'package:webradiooasis/presentation/screens/church_main_screen.dart';
+import 'package:webradiooasis/presentation/providers/audio_player_provider.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
